Extract pure layout viewer helper functions
